Create index.html
MCGA Coin Description:
MCGA Coin is our answer to Trump Coin and a meme coin that will unite Canadians all around the world, showing everyone that together we are ready to withstand any challenge.

Inspired by the spirit of unity, freedom, and the fight for justice, MCGA Coin represents the strength of the Canadian nation and its people. We’re creating a token that’s not just part of the blockchain community but a true movement that brings together those who believe in the values of liberty, equality, and fraternity.

Blockchain: BNB Chain (BSC)

Ticker: MCGA

Total Supply: 41,528,680 MCGA (representing the population of Canada)

Goal: To create a platform that allows Canadians worldwide to unite, support one another, and develop the project further.

MCGA Coin is a coin that keeps up with the times, and in the future, it will be an important symbol for all who are proud of their country and its history. Together, we can achieve more!